Modify, add folder view & move page views
