+++ src/components/views/auth/VerifyAccount/VerifyAccount.jsx
